+++ RELEASE-NOTES.md
@@ -24,6 +24,7 @@
 * #538 "agrest-jaxrs3" module 
 * #541 Remove deprecated metadata responses
 * #557 Upgrade Jackson to 2.13.2.x
+* #558 JacksonService swallows parse exception cause
 
 ## Release 4.10